Add core FastSLAM 1.0 functionality, including particle resampling, and preparing for O(log(n)) particle copying

+++ src/test/kotlin/ca/joelathiessen/kaly2/tests/pc/unit/sensor/SpinnerTest.kt
@@ -66,7 +66,7 @@ class SpinnerTest {
 
         `when`(motor.tachoCount).thenReturn(77)
 
-        assertEquals(spinner.angle.toDouble(), Math.toRadians(77.0), 0.0001)
+        assertEquals(spinner.angle, Math.toRadians(77.0), 0.0001)
     }
 
 }
